Reorg profile folders into react-tree like, profileheadercontainer refactoring, headerImageSelector funtion, HeaderImage abstraction, profileHeader etc.
