+++ src/components/Profile/ProfilePageSections/ProfileHeaderContainer/HeaderInformation/index.tsx
@@ -1,14 +1,14 @@
 import React, { useState } from "react";
 import { View, Text, TouchableOpacity } from "react-native";
 import { AntDesign } from "@expo/vector-icons";
-import { SharePageContainer } from "../../SharePage";
-import { styles } from "./ProfileNameStyles";
-import { verificationStatusEnum } from "../../../services/verificationService/verificationService";
-import { midGrey } from "../../../themeColors";
+import { SharePageContainer } from "../../../../SharePage";
+import { styles } from "./HeaderInformationStyles";
+import { verificationStatusEnum } from "../../../../../services/verificationService/verificationService";
+import { midGrey } from "../../../../../themeColors";
 import { NameFin } from "./NameFin";
-import { profileTypeEnum } from "../profileTypeEnum";
+import { profileTypeEnum } from "../../../profileTypeEnum";
 
-interface ProfileNameProps {
+interface HeaderInformationProps {
   status: number;
   photo: string;
   fin: string;
@@ -17,14 +17,14 @@ interface ProfileNameProps {
   profileSelected: number;
 }
 
-export const ProfileName = ({
+export const HeaderInformation = ({
   status,
   photo,
   fin,
   name,
   workpassType,
   profileSelected
-}: ProfileNameProps) => {
+}: HeaderInformationProps) => {
   const [isDialogVisible, setVisible] = useState(false);
   const showQrCodeBool =
     workpassType === profileTypeEnum.STORED &&
